backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
numpy==1.26.4
Pillow==10.4.0
opencv-python==4.10.0.84
librosa==0.10.2.post1
soundfile==0.12.1
scikit-image==0.24.0
pytest==8.3.2
